+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add `is_bot` field to `gha-utils metadata` output to detect if the current run is not triggered by a human.
 - Check that `gha-utils` CLI can be run as a Python module and with `uv run` and `uvx`.
 
 ## [4.19.1 (2025-09-25)](https://github.com/kdeldycke/workflows/compare/v4.19.0...v4.19.1)